Remove script run.log.
I have no idea WHY but pipping out to tee and appending to a log file to
capture any failed runs of the scripts was causing the install script
to bail out after configuring Unbound.

Any of the commands from unbound-anchor onwards did not execute. I will
need to look into why but it is confirmed to be fixed now that the the
pipe is removed.