Revert "Remove script run.log."
This reverts commit 9a69db47dfcffa5a1c5f1c9da095f3a773f8d979.

It turns out this was not causing any issues. I noticed the script was
exiting either before or after unbound-anchor. I read the man page of
unbound-anchor and it turns out it exits with a return code of 1 in a
successful state.

EXIT CODE
       This tool exits with value 1 if the root anchor was updated using the certificate or if the builtin root-anchor was used.  It exits with code 0 if no update was necessary, if the update was possi‐
       ble with RFC5011 tracking, or if an error occurred.

       You can check the exit value in this manner:
            unbound-anchor -a "root.key" || logger "Please check root.key"
       Or something more suitable for your operational environment.

I set the errexit option at the top of the script so this is what was
triggering it. The script will need to be adjusted to account for this.