Adjust unbound-anchor command so it does not cause script to exit early.
unbound-anchor exits with a return code of 0 or 1. The reasons it uses a
return code of 1 include a successful case as well as a failure case.

This means if it returns 1 you need to log that and handle it otherwise
any script that has the bash option of set -o errexit will exit.