On Centos based distros only update repo cache, not packages. Only add unbound user if it is not present.
I am so use to operating on a Ubuntu/Debian based system that I forgot
the yum packager when invoked with update will actually bump the
versions of packages installed.

I want to avoid this as if I installed a piece of software and it bumped
all of my OS packages and broke something I would be very unpleased
(hello glibc).

Also on Fedora 36 the install script broke due to the unbound user being
added by a build script when building Unbound from source.

This doesn't seem to happen on Ubuntu or Centos or Fedora 33. I have
added a conditional check to only add the user if it is not already
present.

Typically I would prefer to be idempontent and just run the add command
every time as it is easier to reason about but I have the errexit option
enabled to fail fast.